+++ tools/userscript/header.template.js
@@ -20,6 +20,8 @@
 // @resource     IMGPlus https://raw.githubusercontent.com/zellreid/xbox-wishlist/main/browser-extension/src/shared/icons/plus.svg
 // @resource     IMGPreorder https://raw.githubusercontent.com/zellreid/xbox-wishlist/main/browser-extension/src/shared/icons/preorder.svg
 // @resource     IMGPlayAnywhere https://raw.githubusercontent.com/zellreid/xbox-wishlist/main/browser-extension/src/shared/icons/play-anywhere.svg
+// @resource     IMGStar https://raw.githubusercontent.com/zellreid/xbox-wishlist/main/browser-extension/src/shared/icons/star.svg
+// @resource     IMGStarFilled https://raw.githubusercontent.com/zellreid/xbox-wishlist/main/browser-extension/src/shared/icons/star-filled.svg
 // @resource     IMGExpand https://raw.githubusercontent.com/zellreid/xbox-wishlist/main/browser-extension/src/shared/icons/expand.svg
 // @resource     IMGCollapse https://raw.githubusercontent.com/zellreid/xbox-wishlist/main/browser-extension/src/shared/icons/collapse.svg
 // @grant        GM_getResourceURL